feat(blog): authorize Comments schedule reloads
Continue the Blog plan with slice 80: hide raw schedule mutation behind a server-owned read-only facade, add a mandatory host authorizer, serialize trigger attempts, and retain a bounded process-local audit record before returning each authorized mutation outcome.

Tests, source verifiers, formatting, Cargo checks, TCP runtime targets, database targets, browser targets, workflows, and CI were intentionally not run by request.
