+++ crates/rustok-blog/contracts/evidence/blog-comments-tcp-delegation-schedule-trigger.json
@@ -1,0 +1,217 @@
+{
+  "schema_version": 1,
+  "module": "blog",
+  "provider": "comments",
+  "surface": "comments_tcp_delegation_schedule_trigger",
+  "status": "source_verified_no_compile",
+  "compile_policy": "not_run_by_request",
+  "runtime_status": "not_run",
+  "generated_from": "crates/rustok-blog/docs/implementation-plan-slice-80.md",
+  "source_contract": {
+    "runtime_facade": "apps/server/src/services/comments_provider_runtime.rs",
+    "schedule_public_facade": "apps/server/src/services/comments_provider_runtime_keyring_schedule.rs",
+    "schedule_historical_owner": "apps/server/src/services/comments_provider_runtime_keyring_schedule_base.rs",
+    "schedule_trigger_owner": "apps/server/src/services/comments_provider_runtime_keyring_schedule_trigger.rs",
+    "schedule_trigger_guard": "apps/server/src/services/comments_provider_runtime_keyring_schedule_trigger_guard.rs",
+    "slice_79_schedule_guard_preserved": "apps/server/src/services/comments_provider_runtime_keyring_schedule_guard.rs"
+  },
+  "mutation_boundary": {
+    "exported_schedule_handle_read_only": true,
+    "public_initial_host_constructor": true,
+    "public_initial_file_constructor": true,
+    "public_status_reads": true,
+    "public_keyring_provider": true,
+    "public_raw_replace_method": false,
+    "public_raw_file_reload_method": false,
+    "raw_mutation_visibility": "server_parent_only",
+    "trigger_required_for_exported_mutation": true,
+    "historical_owner_blob_preserved": true
+  },
+  "trigger_composition": {
+    "programmatic_only": true,
+    "mandatory_authorizer": true,
+    "permissive_default_authorizer": false,
+    "standalone_handle_and_trigger_rejected": true,
+    "trigger_owned_handle_inserted_before_slice_79_guard": true,
+    "slice_79_runtime_policy_guard_reused": true,
+    "http_route": false,
+    "graphql_route": false,
+    "native_rpc_route": false,
+    "mcp_tool": false,
+    "cli_command": false,
+    "signal_handler": false,
+    "watcher": false,
+    "poller": false
+  },
+  "request_context": {
+    "non_nil_request_uuid": true,
+    "non_nil_actor_uuid": true,
+    "typed_principal_kind": "rustok_api::AuthPrincipalKind",
+    "raw_bearer_token": false,
+    "raw_session_token": false,
+    "raw_oauth_claims": false,
+    "caller_supplied_roles": false,
+    "caller_supplied_audit_actor_override": false
+  },
+  "authorization": {
+    "delegated_user_always_ineligible": true,
+    "delegated_user_rejected_before_authorizer": true,
+    "direct_user_requires_authorizer": true,
+    "service_requires_authorizer": true,
+    "request_fields": [
+      "request_id",
+      "actor_id",
+      "principal_kind",
+      "operation",
+      "source_category",
+      "current_generation",
+      "requested_generation"
+    ],
+    "closed_errors": [
+      "denied",
+      "unavailable"
+    ],
+    "free_form_reason_retained": false,
+    "authorizer_called_under_operation_serialization": true,
+    "authorizer_reentrancy_supported": false
+  },
+  "serialization": {
+    "one_process_local_operation_mutex": true,
+    "audit_clock_checked_before_mutation": true,
+    "audit_mutex_checked_before_mutation": true,
+    "audit_sequence_checked_before_mutation": true,
+    "audit_guard_held_during_mutation": true,
+    "final_outcome_appended_before_return": true,
+    "successful_return_without_local_record": false,
+    "authorization_failure_invokes_mutation": false,
+    "mutation_failure_preserves_slice_79_snapshot": true
+  },
+  "audit": {
+    "process_local": true,
+    "append_only_within_retained_window": true,
+    "minimum_capacity": 1,
+    "default_capacity": 256,
+    "maximum_capacity": 1024,
+    "oldest_evicted_at_capacity": true,
+    "checked_monotonic_sequence": true,
+    "sequence_reused": false,
+    "sequence_overflow_fails_before_mutation": true,
+    "outcomes": [
+      "preflight_rejected",
+      "principal_ineligible",
+      "authorization_denied",
+      "authorization_unavailable",
+      "replacement_rejected",
+      "replacement_succeeded"
+    ],
+    "record_fields": [
+      "sequence",
+      "occurred_at_unix_ms",
+      "request_id",
+      "actor_id",
+      "principal_kind",
+      "operation",
+      "outcome",
+      "source_category",
+      "previous_generation",
+      "requested_generation",
+      "current_generation"
+    ],
+    "file_path": false,
+    "key_ids": false,
+    "secret_values": false,
+    "schedule_json": false,
+    "credential": false,
+    "delegation_nonce": false,
+    "raw_mutation_error": false,
+    "token_values": false,
+    "arbitrary_authorizer_metadata": false,
+    "external_sink": false,
+    "durable": false,
+    "database_backed": false,
+    "broker_backed": false,
+    "restart_safe": false
+  },
+  "debug_redaction": {
+    "schedule_handle_redacted_by_existing_debug": true,
+    "authorizer_details": false,
+    "audit_capacity": true,
+    "audit_record_count": true,
+    "audit_actor_ids": false,
+    "audit_request_ids": false,
+    "paths": false,
+    "key_ids": false,
+    "secrets": false
+  },
+  "preserved_behavior": {
+    "schedule_schema_version": 2,
+    "activation_policy_changed": false,
+    "retirement_policy_changed": false,
+    "overlap_formula_changed": false,
+    "clock_floor_changed": false,
+    "generation_policy_changed": false,
+    "replay_gate_changed": false,
+    "delegation_scheme_changed": false,
+    "delegation_version_changed": false,
+    "key_id_hmac_binding_changed": false,
+    "service_bearer_reads": true,
+    "system_moderation": true,
+    "authority_override_precedence": true,
+    "channel_override_precedence": true,
+    "provider_override_precedence": true,
+    "loopback_only": true,
+    "framing_changed": false,
+    "deadline_changed": false,
+    "listener_lifecycle_changed": false,
+    "shutdown_changed": false
+  },
+  "dependency_contract": {
+    "new_direct_dependencies": [],
+    "manifest_changed": false,
+    "cargo_lock_changed": false,
+    "feature_changed": false
+  },
+  "non_claims": [
+    "http_trigger",
+    "graphql_trigger",
+    "native_rpc_trigger",
+    "mcp_trigger",
+    "cli_trigger",
+    "automatic_file_watch",
+    "polling",
+    "signal_reload",
+    "durable_audit",
+    "database_audit",
+    "broker_audit",
+    "transactional_outbox",
+    "persisted_generation",
+    "persisted_schedule_digest",
+    "restart_rollback_prevention",
+    "cross_process_audit_sequence",
+    "cross_process_clock_synchronization",
+    "distributed_atomic_activation",
+    "cloud_secret_manager_sdk",
+    "kms_hsm",
+    "shared_replay_store",
+    "durable_replay_store",
+    "multi_replica_replay_prevention",
+    "restart_safe_replay_prevention",
+    "secret_zeroization",
+    "locked_memory",
+    "file_permission_attestation",
+    "tls_mtls",
+    "non_loopback_publication",
+    "runtime_execution"
+  ],
+  "execution": {
+    "rust_tests_run": false,
+    "javascript_verifiers_run": false,
+    "cargo_commands_run": false,
+    "formatting_run": false,
+    "tcp_runtime_run": false,
+    "database_run": false,
+    "browser_run": false,
+    "workflow_run": false,
+    "ci_run": false
+  }
+}
